Fix: 예약 기능 로직 변경 및 예약 상세 조회 추가
신정훈 님
1. 타 사용자 예약 기능 사용 불가
2. 예약 상세 조회 기능 추가 (router, 아키텍처 변경)
3. 예약 기능 등록 및 수정 예외 처리 보완

+++ controllers/reservation.controller.js
@@ -1,5 +1,4 @@
 const ReservationService = require('../services/reservation.service.js');
-// const moment = require('moment');
 
 class ReservationController {
   constructor() {
@@ -12,14 +11,13 @@ class ReservationController {
     const userId = res.locals.user;
 
     const { guestId, sitterId } = req.params;
-    const { status, message, resulveService } =
-      await this.reservationService.createReservation(
-        startDateTime,
-        endDateTime,
-        guestId,
-        sitterId,
-        userId
-      );
+    const { status, message } = await this.reservationService.createReservation(
+      startDateTime,
+      endDateTime,
+      guestId,
+      sitterId,
+      userId
+    );
 
     res.status(status).json({ message });
   };
@@ -29,37 +27,41 @@ class ReservationController {
     const { reservationsId } = req.params;
     const userId = req.userId;
 
-    const { status, message } = await this.reservationService.deleteResulve(
-      reservationsId,
-      userId
-    );
+    const { status, message } = await this.reservationService.deleteResulve(reservationsId, userId);
 
     res.status(status).json({ message });
   };
 
   // 예약 전체 조회
   getresulve = async (req, res, next) => {
-    const { status, messge, reservationService } =
-      await this.reservationService.getResulve();
+    const { status, messge, reservationService } = await this.reservationService.getResulve();
 
     res.status(status).json({ messge, reservationService });
+  };
+
+  // 예약 상세 조회
+  getreservation = async (req, res, next) => {
+    const { reservationsId } = req.params;
+
+    const { status, message } = await this.reservationService.getreservation(reservationsId);
+
+    res.status(status).json({ message, reservationsId });
   };
 
   // 예약 수정
 
   putresulve = async (req, res, next) => {
     const { startDateTime, endDateTime } = req.body;
-    const userId = res.locals.user;
+    const userId = req.userId;
 
     const { reservationsId } = req.params;
 
-    const { status, message, resulveService } =
-      await this.reservationService.putReservation(
-        startDateTime,
-        endDateTime,
-        reservationsId,
-        userId
-      );
+    const { status, message } = await this.reservationService.putReservation(
+      startDateTime,
+      endDateTime,
+      reservationsId,
+      userId
+    );
 
     res.status(status).json(message);
   };
